Updates: 10GB 4:32

+++ Common/FileCleaner.cs
@@ -12,10 +12,16 @@
         public static void CleanFiles()
         {
             files.Remove(Options.SortedOutputDataFile); //leave output file
-            foreach (string fileName in files)
+
+            int filesCount = files.Count;
+            if (filesCount == 0)
+                return;
+
+            foreach (string fileName in files.ToList())
             {
                 File.Delete(fileName);
             }
+            Logger.Log($"Deleted: {filesCount} temp files successfully.");
         }
     }
 }
